build(angular): Use the Angular compiler to compile the @sentry/angular SDK (WEB-622)
Switch to the Angular compiler (via Angular CLI) to ensure that all Angular 1st-class
support items of the SDK are compiled in a way that the library are compatible with
Angular. This fixes the compilation errors when using `TraceDirective` and `TraceModule`
that are currently thrown on application compilation.
Uses the Angular 10 compiler which ensures compatibility with Angular 10-13 apps.

FIXES GH-3282
FIXES WEB-622